feat: add useScrambledText hook
